Remove extra blank lines to avoid confusion (#78)
Because of so many blank lines, I could not see the lower part of code and got confused if this config file is correct or what. After sometime, saw that there were blank lines.

So suggest to remove these blank lines so other developers won't get confused as I did.